+++ src/packages/CSS/Controls/GetFeatureInfo/DSFRgetFeatureInfoStyle_ex.css
@@ -1,0 +1,15 @@
+/* GET FEATURE INFO */
+
+.gpf-btn-icon-getfeatureinfo {
+    background-image: url("img/NoGFI.svg");
+    background-repeat: no-repeat;
+    background-size: auto auto;
+    background-position: center center;
+}
+
+button[id^="GPgetFeatureInfoPicto-"][aria-pressed="true"] {
+    background-image: url("img/GFI.svg");
+    background-repeat: no-repeat;
+    background-size: auto auto;
+    background-position: center center;
+}
