--- src/packages/CSS/Controls/GetFeatureInfo/GPFgetFeatureInfoStyle.css
@@ -1,11 +1,0 @@
-/* GET FEATURE INFO */
-
-.GPgetFeatureInfoPicto {
-    background-position: -25px center;
-    background-image: url("img/GPactivateGfi.png");
-    background-repeat: no-repeat;
-}
-
-button[id^="GPgetFeatureInfoPicto-"][aria-pressed="true"] {
-    background-position: 1px center;
-}
