+++ src/packages/Controls/GetFeatureInfo/GetFeatureInfoDOM_ex.js
@@ -1,0 +1,62 @@
+var GetFeatureInfoDOM = {
+
+    /**
+    * Add uuid to the tag ID
+    * @param {String} id - id selector
+    * @returns {String} uid - id selector with an unique id
+    */
+    _addUID : function (id) {
+        var uid = (this._uid) ? id + "-" + this._uid : id;
+        return uid;
+    },
+
+    /**
+     * Main container (DOM)
+     *
+     * @returns {DOMElement} DOM element
+     */
+    _createMainContainerElement : function () {
+        var container = document.createElement("div");
+        container.id = this._addUID("GPgetFeatureInfo");
+        container.className = "GPwidget gpf-widget gpf-widget-button";
+        return container;
+    },
+
+    // ################################################################### //
+    // ################# Methods to display Main Panel ################### //
+    // ################################################################### //
+
+    /**
+     * Creation du container du picto du controle (DOM)
+     * @param {Boolean} active - active
+     * @returns {DOMElement} DOM element
+     */
+    _createMainPictoElement : function (active) {
+        var self = this;
+
+        var button = document.createElement("button");
+        button.id = this._addUID("GPgetFeatureInfoPicto");
+        button.className = "GPshowOpen GPshowAdvancedToolPicto GPgetFeatureInfoPicto gpf-btn gpf-btn-icon gpf-btn-icon-getfeatureinfo fr-btn";
+        button.title = "activer/desactiver l'interrogation des couches";
+        button.setAttribute("tabindex", "0");
+        button.setAttribute("aria-pressed", active);
+        button.setAttribute("type", "button");
+ 
+        if (button.addEventListener) {
+            button.addEventListener("click", function (e) {
+                var status = (e.target.ariaPressed === "true");
+                e.target.setAttribute("aria-pressed", !status);
+                self.onActivateGetFeatureInfoElementChange(e);
+            });
+        } else if (button.attachEvent) {
+            button.attachEvent("onclick", function (e) {
+                var status = (e.target.ariaPressed === "true");
+                e.target.setAttribute("aria-pressed", !status);
+                self.onActivateGetFeatureInfoElementChange(e);
+            });
+        }
+        return button;
+    }
+};
+
+export default GetFeatureInfoDOM;
